refactor : Student Entity 수정 학교명 추가
